первые два замечания
